Fix memory leak
When `NextToken` creates one token, the token is used to build the DOM
Tree. After this token is used, the token must be initalized to create
the next token.

In the past, `release` was called to initialize the token. This causes a
memory leak. In this patch, the token is initalized by calling `reset`.

Also added logic to generate kEndOfFile token when end of file is
reached during parsing.